Added course descriptions+prereqs to catalogscraper
