+++ scraper/catalogscraper.py
@@ -1,11 +1,24 @@
 import requests
+from selenium import webdriver
+from selenium.webdriver.common.by import By
+from selenium.webdriver.chrome.service import Service
+from webdriver_manager.chrome import ChromeDriverManager
 from bs4 import BeautifulSoup
 from unicodedata import normalize
+import time
+
+class Course:
+    def __init__(self):
+        self.title = ""
+        self.full_description = ""
+        self.prereqs = ""
+        self.coreqs = ""
 
 class CatalogData:
     def __init__(self):
         self.terms = []
         self.milestones = []
+        self.all_courses = {}
 
 # MANUALLY GO THROUGH THIS FUNCTION AND CHANGE URLS/YEARS TO UPDATE
 def getURLs():
@@ -38,29 +51,107 @@ def getURLs():
 
     return urls
 
+def clickAllLinks(driver: webdriver.Chrome):
+    # All of the classes that CS, CE, or EE takes
+    class_prefixes = ['COSC', 'ECE', 'EE', 'MATH', 'ENGL', 'BIOL', 'CHEM', 'PHYS', 'EF']
 
+    # Grab all the a elements that have hrefs
+    links = driver.find_elements(By.TAG_NAME, 'a')
+    
+    # go through all the grabbed elements and look for classes we should click on
+    for link in links:
+        # wrap in a try except block because we get some non-class elements that throw exceptions
+        try:
+            if link.text.split(' ')[0] in class_prefixes:
+                # these print comments are for debugging
+                # print(link.text, end='  ')
+                # print(link.location)
+                link.click()
+                time.sleep(0.5)
+                link.click()
+                time.sleep(0.5)
+
+        except Exception as e:
+            pass
+    
+    return driver.page_source
+
+def getAllCourses(soup: BeautifulSoup):
+    # prefixes of classes taken for this major
+    class_prefixes = ['COSC', 'ECE', 'EE', 'MATH', 'ENGL', 'BIOL', 'CHEM', 'PHYS', 'EF']
+
+    # isolate all div elements that contain the box that pops up when you click on a class
+    class_descriptions = soup.find_all("div")
+    class_descriptions = [x for x in class_descriptions if ('<h3>' in str(x)) and (x.h3.text.strip().split(' ')[0] in class_prefixes) and 'facebook' not in str(x)]
+
+    # create 'course' class for each course
+    all_courses = {}
+    for c in class_descriptions:
+        temp_course = Course()
+        text = " ".join(c.text.strip().split())
+        text = text.replace(' *', '')
+        text = text.replace('*', '')
+        text = text.replace(' and ', '+')
+        text = text.replace(', or', ' or')
+        text = text.replace(', ', ' or ')
+        text = text.replace('  ', ' ')
+
+        # get name of class
+        temp_course.title = text.split()[0] + ' ' + text.split()[1]
+
+        # put in full description
+        temp_course.full_description = text
+
+        # find prereqs if there are any
+        index_prereq = text.find('Prerequisite(s):')
+        if index_prereq != -1:
+            prereq_end = text[index_prereq:].find('.')
+            temp_course.prereqs = text[index_prereq+17:index_prereq+prereq_end]
+        
+        # find coreqs if there are any
+        index_coreq = text.find('Corequisite(s):')
+        if index_coreq != -1:
+            coreq_end = text[index_coreq:].find('.')
+            temp_course.coreqs = text[index_coreq+16:index_coreq+coreq_end]
+
+        all_courses[temp_course.title] = temp_course
+
+    return all_courses
 
 def main():
     urls = getURLs()
 
-    catalogs = {} # this will be a dict of dicts where the key is major-year, and value is a dict where those keys are terms and the value is a list of classes
+    catalogs = {} # this will be a dict of dicts where the key is major-year, and value is the CatalogData class
+
+    # set up the driver options
+    options = webdriver.ChromeOptions()
+    options.add_argument('--ignore-certificate-errors')
+    options.add_argument('--incognito')
+    options.add_argument('--headless')
+
+    webdriver_service = Service(ChromeDriverManager().install()) 
+    # Choose Chrome Browser
+    driver = webdriver.Chrome(service=webdriver_service, options=options)
 
     # loop through every URL to grab catalog
     for major, majorURLs in urls.items():
         for year, url in majorURLs.items():
-            print(year)
+            print(f'{major}-{year}')
 
             # keep up with the term number
             term = 0
             try:
-                # grab the page and let bs4 convert it
-                page = requests.get(url)
-                soup = BeautifulSoup(page.content, 'html.parser')
+                # grab the page, click all the links to get everything on the DOM, then let BS4 parse it
+                driver.get(url)
+                page = clickAllLinks(driver)
+                soup = BeautifulSoup(page, 'html.parser')
 
-                # f = open('test.txt', 'w')
-                # f.write(soup.prettify())
-                # f.close()
-                # break
+                # set the value of a certain catalog to an empty dict so we can fill it in later
+                catalogs[f'{major}-{year}'] = CatalogData()
+
+                # Before we grab the schedule, we can go ahead and store all of the courses for this catalog
+                all_courses = getAllCourses(soup)
+                catalogs[f'{major}-{year}'].all_courses = all_courses.copy()
 
                 # the table we want has a class of 'acalog-core'
                 # we want the first table with this class
@@ -70,9 +161,6 @@ def main():
 
                 # temporary list for storing classes needed in a term
                 temp_classes = []
-
-                # set the value of a certain catalog to an empty dict so we can fill it in later
-                catalogs[f'{major}-{year}'] = CatalogData()
                 for tr in trs:
                     tds = tr.find_all('td')
 
@@ -115,17 +203,29 @@ def main():
             except:
                 print('unable to access page\n')
 
-        #     break
-        # exit()
 
+    # open files for writing
+    file_schedule = open('catalog-schedules.txt', 'w')
+    file_courses = open('catalog-courses.txt', 'w')
     for catalog, catalog_class in catalogs.items():
-        print(f'{catalog}: ')
+        file_schedule.write(f'{catalog}: \n')
         for i, courses in enumerate(catalog_class.terms):
-            print(f'Term {i+1}:\nMilestones: {catalog_class.milestones[i]}\n')
+            file_schedule.write(f'Term {i+1}:\nMilestones: {catalog_class.milestones[i]}\n\n')
             for course in courses:
-                print(course)
-            print('')
-        print('\n\n*****************************************************************\n\n')
+                file_schedule.write(course)
+                file_schedule.write('\n')
+            file_schedule.write('\n')
+        
+        file_courses.write(f'{catalog}: \n')
+        for name, course in catalog_class.all_courses.items():
+            file_courses.write(f'{name}\n')
+            file_courses.write(f'Description: {course.full_description}\n')
+            file_courses.write(f'Prerequisites: {course.prereqs}\n')
+            file_courses.write(f'Corequisites: {course.coreqs}\n')
+            file_courses.write('\n')
+
+    file_schedule.close()
+    file_courses.close()
         
 
 if __name__ == '__main__':
